Develop fake

+++ database/migrations/2025_07_21_144435_create_products_table.php
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->foreignId('default_cogs_account_id')->nullable()->constrained('accounts');
             $table->foreignId('default_stock_input_account_id')->nullable()->constrained('accounts');
             $table->foreignId('default_price_difference_account_id')->nullable()->constrained('accounts');
-            $table->decimal('average_cost', 15, 4)->default(0.00);
+            $table->unsignedBigInteger('average_cost')->default(0);
             $table->integer('quantity_on_hand')->default(0);
 
             $table->softDeletes();
